Fix test by increasing precision of y_star_at to f64
Not the ideal fix. Need to take a better look at this later.

+++ src/main.rs
@@ -868,19 +868,19 @@ impl Bisector {
     fn y_star_at(&self, sites: &[Point], x: f32) -> f32 {
         let (a, b) = self.ab(sites);
 
-        let dx = b.pos.x - a.pos.x;
-        let dy = b.pos.y - a.pos.y;
-        let x = x - a.pos.x;
+        let dx = b.pos.x as f64 - a.pos.x as f64;
+        let dy = b.pos.y as f64 - a.pos.y as f64;
+        let x = x as f64 - a.pos.x as f64;
 
         let x2 = x * x;
         let dx2 = dx * dx;
         let dy2 = dy * dy;
 
-        let sqrt = |x: f32| x.sqrt();
+        let sqrt = |x: f64| x.sqrt();
 
-        a.pos.y
+        (a.pos.y as f64
             + sqrt(x2 + (dx2 - 2.0 * dx * x + dy2).powi(2) / (4.0 * dy2))
-            + (dx2 - 2.0 * dx * x + dy2) / (2.0 * dy)
+            + (dx2 - 2.0 * dx * x + dy2) / (2.0 * dy)) as f32
     }
 
     /// Return if point is on the left side or right side of the hyperbola, obtained by the
